Use dedicated entities to group code logic
This improves readability and maintainability. It also was a good test
case to try out some of the techniques in the "Refactoring" book.

As I'm not familiar with the Python setup for TDD and this script is
quite managable, I went at it without focussing too much on it unit
tests at this time.